# apps/backend/internal/handler/handler.go
package handler

import (
	"encoding/json"
	"net/http"
	"time"

	"github.com/go-chi/chi/v5"
	"gorm.io/gorm"

	"github.com/react-go-monorepo/backend/internal/config"
)

// Handler holds the dependencies for the HTTP handlers
type Handler struct {
	db  *gorm.DB
	cfg *config.Config
}

// New creates a new handler instance
func New(db *gorm.DB, cfg *config.Config) *Handler {
	return &Handler{
		db:  db,
		cfg: cfg,
	}
}

// HealthCheck handles the health check endpoint
func (h *Handler) HealthCheck(w http.ResponseWriter, r *http.Request) {
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"status": "ok",
		"time":   time.Now().Format(time.RFC3339),
	})
}

// Login handles user login
func (h *Handler) Login(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement login logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{
		"error": "Login not implemented yet",
	})
}

// RefreshToken handles token refresh
func (h *Handler) RefreshToken(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement refresh token logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNotImplemented)
	json.NewEncoder(w).Encode(map[string]string{
		"error": "Refresh token not implemented yet",
	})
}

// AuthMiddleware is a middleware for authenticating requests
func (h *Handler) AuthMiddleware(next http.Handler) http.Handler {
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		// TODO: Implement JWT validation
		// For now, just pass through
		next.ServeHTTP(w, r)
	})
}

// GetTodos handles GET /todos
func (h *Handler) GetTodos(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement get todos logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode([]map[string]interface{}{})
}

// CreateTodo handles POST /todos
func (h *Handler) CreateTodo(w http.ResponseWriter, r *http.Request) {
	// TODO: Implement create todo logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusCreated)
	json.NewEncoder(w).Encode(map[string]string{
		"message": "Todo created",
	})
}

// GetTodoById handles GET /todos/{todoId}
func (h *Handler) GetTodoById(w http.ResponseWriter, r *http.Request) {
	todoID := chi.URLParam(r, "todoId")
	// TODO: Implement get todo by id logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"id":      todoID,
		"message": "Get todo by id not implemented yet",
	})
}

// UpdateTodo handles PUT /todos/{todoId}
func (h *Handler) UpdateTodo(w http.ResponseWriter, r *http.Request) {
	todoID := chi.URLParam(r, "todoId")
	// TODO: Implement update todo logic
	w.Header().Set("Content-Type", "application/json")
	json.NewEncoder(w).Encode(map[string]string{
		"id":      todoID,
		"message": "Update todo not implemented yet",
	})
}

// DeleteTodo handles DELETE /todos/{todoId}
func (h *Handler) DeleteTodo(w http.ResponseWriter, r *http.Request) {
	_ = chi.URLParam(r, "todoId") // TODO: Use todoID when implementing delete logic
	// TODO: Implement delete todo logic
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(http.StatusNoContent)
}